Add files via upload
This is an Online event management system that serves the functionality of an event manager. It allows the user to select from a list of event types. Once the user enters an event type and the system then allows the user to select the date and time of event, place. All this data is logged in the database and the user is given a receipt number for his booking. This data is then sent to the administrator  and they may interact with the client as per his requirements.